파일 convert 메소드 분리
